fix: added support of expressions inside classname and style names
