+++ apps/kitchen-sink/src/App.tsx
@@ -1,7 +1,7 @@
 import type { ApplicationDTO } from "@ensembleui/react-framework";
 import { EnsembleApp } from "@ensembleui/react-runtime";
 // Screens
-// import MenuYAML from "./ensemble/screens/menu.yaml";
+import MenuYAML from "./ensemble/screens/menu.yaml";
 import HomeYAML from "./ensemble/screens/home.yaml";
 import WidgetsYAML from "./ensemble/screens/widgets.yaml";
 import LayoutsYAML from "./ensemble/screens/layouts.yaml";
@@ -60,6 +60,11 @@ const testApp: ApplicationDTO = {
     },
   ],
   screens: [
+    {
+      id: "menu",
+      name: "Menu",
+      content: String(MenuYAML),
+    },
     {
       id: "home",
       name: "Home",
